CITATION.cff: remove version and date-released and use canonical
An issue with the CITATION.cff generated via doi2cff is that it is tied to a
given tag and that the update process is unclear (and cyclic). Until we get
this figured out, the non mandatory version and date-released attributes are
removed and the doi is adjusted to use the Zenodo DOI for all versions which
always resolves to the latest version.